python-requirements.txt
Django==1.7.5
django-annoying==0.8.1
django-extensions==1.5.0
django-report-tools==0.2.2
django-shell-plus==1.1.5
djangorestframework==3.0.5
ipython==3.0.0
MySQL-python==1.2.5
six==1.9.0
